Refactor coordination classes and relocate `LandLocation`.
Decouple `LandLocation` from `coord.py` and move it to `land.py` with improved modularity. Replaced `dataclass` with an explicit class-based approach to enhance flexibility and maintainability. Enhanced `LandLocation` initialization to support name attribute and maintain original coordinates.
